requirements-blob.txt
fastapi==0.115.5
uvicorn==0.32.1
gunicorn==23.0.0
python-multipart==0.0.20
openai==1.57.2
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
tenacity==9.0.0
docling==2.13.0
azure-storage-blob==12.19.0
azure-identity==1.15.0